Fix function authenticateJWT and Add function checkRoleQuery

+++ routes/user.js
@@ -3,11 +3,11 @@ const userCtrl = require('../controllers').controller.User;
 const authenticate = require('../middleware/authenticate');
 const role = require('../middleware/role');
 
-userRoute.get('/s', authenticate.authenticateJWT, userCtrl.getMany);
+userRoute.get('/s', role.checkRoleQuery(["name", "email"], ["Admin"], { read: true }), userCtrl.getMany);
 userRoute.get('/:id', authenticate.authenticateJWT, role.checkRole(["Admin"], { read: true }), userCtrl.getById);
 userRoute.post('/', userCtrl.createData);
-userRoute.put('/:id', authenticate.authenticateJWT, userCtrl.updateById);
-userRoute.delete('/:id', authenticate.authenticateJWT, userCtrl.deleteById);
+userRoute.put('/:id', authenticate.authenticateJWT, role.checkRole(["Admin", "Customer"], { update: true }), userCtrl.updateById);
+userRoute.delete('/:id', authenticate.authenticateJWT, role.checkRole(["Admin"], { delete: true }), userCtrl.deleteById);
 userRoute.post('/login', authenticate.getToken);
 
 module.exports = userRoute;
